Clustering1.js updated
It should work now, tweeked some number and it works a lot of the time,
but sometimes throws up a bug (in chrome at least) about chrome devtools
discconecting, causing the page to crash. don't know if this is a chrome
issue, or a my computer issue, but just reducing the multiplyer for the
amount of tweets hard clustered can fix it in the short run (on the 3rd
last line i think). the error happens in the hard clustering, possibly
because it uses too much ram or something? (seems to be a known chrome
bug with no fix as of yet :/